* Cleanup in utils. * Isolated edge calculation to its own file.
